Public API Reference

This is an exhaustive list of all exported constants, types and functions in CompositeTypes.jl.

Constants

Functions

CompositeTypes.component — Method

component(x, I...)

The component of x that corresponds to the given index.

CompositeTypes.components — Method

components(x)

The components of the composite object x.

CompositeTypes.iscomposite — Method

iscomposite(x)

Is x a composite object?

If an object is composite, then its components can be queried using the components function.

CompositeTypes.ncomponents — Method

ncomponents(x)

The number of components of composite object x.

CompositeTypes.setcomponent! — Method

setcomponent!(x, v, I...)

Set the component of x that corresponds to the given index to v.

Types