Indexing

Public API Reference

This is an exhaustive list of all exported constants, types and functions in CompositeTypes.Indexing.

Constants

Functions

Types

CompositeTypes.Indexing.Component — Type

Component{I}

Component index of a composite type.

Internally, this index simply stores an index or a range of indices to components.

CompositeTypes.Indexing.ComponentIndex — Type

ComponentIndex{I}

Supertype of indices that select components of an object rather than its entries.

Internal API Reference

This is an exhaustive list of all non-exported constants, types and functions in CompositeTypes.Indexing.

Warning

Unexported functions and types are subject to change across different releases of the package, even if the release is said to be non-breaking.

Constants

Functions

CompositeTypes.Indexing.similarindex — Method

similarindex(idx::Component, I)

Create a new component index of the same type but with a different value.

CompositeTypes.Indexing.to_index — Method

to_index(idx::Component)

Convert the component index to an index of components.

Types